feat(shared-api): try sharing member response schemas
Move member response schemas into the shared package and point the
functions APIs at that shared source. Preserve this failed experiment
so we can revisit what broke in Elysia route inference later.
