+++ functions/src/admin-messages-api/CHANGELOG.md
@@ -1,4 +1,4 @@
-##  (2026-03-14)
+##  (2026-03-15)
 
 ### Features
