+++ functions/src/members-api/CHANGELOG.md
@@ -1,4 +1,4 @@
-##  (2026-03-13)
+##  (2026-03-15)
 
 ### Features
 
@@ -20,4 +20,5 @@
 
 * add typed handleRequest wrapper to resolve 898 lint errors in test files ([e9015f4](https://github.com/markgoho/doula-cooperative/commit/e9015f43e8a885496f4a658465d50ed3756c08d4))
 * make verify-email endpoint idempotent for already-verified emails ([#68](https://github.com/markgoho/doula-cooperative/issues/68)) ([9de2431](https://github.com/markgoho/doula-cooperative/commit/9de2431549cd02ac90a4e69515c9bd41fa33c94e))
+* **profile-webhook-api:** bcc webmaster on profile update emails ([9774f4a](https://github.com/markgoho/doula-cooperative/commit/9774f4a557aa281fa7ff45ee5f77084b0be5f4c9))
 * remove unnecessary type assertions in test files ([f774f3b](https://github.com/markgoho/doula-cooperative/commit/f774f3b3cab00ca2afc3e24d0f6ff50e52cd6ff6))
