+++ functions/tsconfig.json
@@ -7,7 +7,11 @@
     "outDir": "lib",
     "skipLibCheck": true,
     "sourceMap": true,
-    "target": "es2022"
+    "target": "es2022",
+    "baseUrl": ".",
+    "paths": {
+      "@doula-coop/shared": ["../shared/src/index.ts"]
+    }
   },
   "compileOnSave": true,
   "include": ["src"],
