+++ package.json
@@ -33,6 +33,7 @@
     "@imagekit/nodejs": "^7.3.0",
     "@types/bun": "^1.3.10",
     "conventional-changelog-cli": "^5.0.0",
+    "elysia": "^1.2.0",
     "eslint": "^9.39.4",
     "eslint-plugin-unicorn": "^63.0.0",
     "firebase-admin": "^13.7.0",
