+++ shared/src/index.ts
@@ -4,5 +4,10 @@ export type { SubscriptionStatus } from "./subscription-status.js";
 export { WELCOME_EMAIL_STATUSES } from "./welcome-email-status.js";
 export type { WelcomeEmailStatus } from "./welcome-email-status.js";
 
-export type { ApiMemberResponse } from "./api-member-response.js";
+export {
+  MemberResponseSchema,
+  SubscriptionStatusSchema,
+  WelcomeEmailStatusSchema,
+} from "./api-member-response-schema.js";
+export type { ApiMemberResponse } from "./api-member-response-schema.js";
 export type { UnlinkedProfile } from "./unlinked-profile.js";
